pyproject.toml
[tool.poetry]
name = "document-color-distribution"
version = "0.1.0"
description = ""
authors = ["diegulio <diego.machado@usach.cl>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
ipykernel = "^6.29.5"
numpy = "^2.2.1"
pillow = "^11.1.0"
pdf2image = "^1.17.0"
streamlit = "^1.41.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
